refactor(scanner): split scalar scanning into a submodule (#109)
scanner.rs 3174 -> 2093; scalar scanning to parser/scanner/scalars.rs. No API change (semver clean). 5217 tests + Miri green.
